Bats avoid ground when searching
The ground avoidance previously implemented for approaching targets is now added for search mode as well, with even larger distances, to account for the generally higher speed of search mode. When approaching, bats now fly away horizontally as well as up when they hit a wall. Furthermore, some more comments.