Separate Surface and Area Damping
Surface damping, Area damping and air friction are now separately implemented, allowing for more realistic droplet properties. (Surface damping is proportional to change in vertex separation and area damping is proportional to change in area.) This leads to an interesting result, as the droplet starts to oscillate in a mode keeps its area and vertex separations approximately constant. However, as change in surface angles does not yet result in friction, this is how the oscillation manifests.

Furthermore, a saturate function was added to extra_math.js.