Add droplet-droplet interaction
Droplet-droplet interaction is now implemented. When two droplets intersect, their vertices and the interaction points get stitched together.

Though the result seems to be correct in most cases. Due to motion some new formed droplets will self-interact and explode. Self-interaction is therefore a logical next step.

As of this point, the newly added intersection points get inserted in the same order every time. It is not clear yet if this will actually give the correct result every time. This should be investigated.